third_party/blink/web_tests/fast/table/calculated-width-on-col-within-colgroup.html

<!DOCTYPE html>
<style>
    col {
        width: calc(-10in - 10%);
    }
</style>
<script src="../../resources/check-layout.js"></script>
<table>
    <colgroup width="200">
        <col data-expected-width=200>
        <col data-expected-width=200>
    </colgroup>
    <tr>
        <td data-expected-width=200></td>
        <td data-expected-width=200></td>
    </tr>
</table>
<script>
    checkLayout('table')
</script>
<p> crbug.com/446936: Use a used value of Auto on col elements that have a calc value. This is permitted by the spec and avoids ASSERTs.</p>
